FinalEvent

+++ ConsommiTounsi/Models/Products/ProductRating.cs
@@ -8,7 +8,7 @@ namespace ConsommiTounsi.Models.Products
     public class ProductRating
     {
         public long id { get; set; }
-        public double starRate { get; set; }
+        public int starRate { get; set; }
         public long visitCount { get; set; }
         public double personalRating { get; set; }
         public double globalRating { get; set; }
